Formating updates.
Fixed couple fo whitespaces.
Restructured condition structure.
Fixed variable casts.

+++ rtos/os/os.h
@@ -33,11 +33,11 @@
 #define NULL                        (0)
 
 /* Number of system ticks per second. */
-#define OS_TICKS_PER_SEC            (uint32_t)(100)
-#define MS_TO_TICK(a)               (((uint64_t)(a) * OS_TICKS_PER_SEC) / (1000))
-#define TICK_TO_MS(a)               (((uint64_t)(a) * 1000) / (OS_TICKS_PER_SEC))
-#define US_TO_HW_TICK(a)            (((uint64_t)(a) * OS_HW_TICKS_PER_SEC) / (1000000))
-#define HW_TICK_TO_US(a)            (((uint64_t)(a) * 1000000) / (OS_HW_TICKS_PER_SEC))
+#define OS_TICKS_PER_SEC            (uint64_t)(100)
+#define MS_TO_TICK(a)               ((uint64_t)(((uint64_t)(a) * OS_TICKS_PER_SEC) / (1000)))
+#define TICK_TO_MS(a)               ((uint64_t)(((uint64_t)(a) * 1000) / (OS_TICKS_PER_SEC)))
+#define US_TO_HW_TICK(a)            ((uint64_t)(((uint64_t)(a) * OS_HW_TICKS_PER_SEC) / (1000000)))
+#define HW_TICK_TO_US(a)            ((uint64_t)(((uint64_t)(a) * 1000000) / (OS_HW_TICKS_PER_SEC)))
 
 /* Some useful macros. */
 #define OFFSETOF(type, field)       ((int) &(((type *) 0)->field))
